find_refs_by_id: Use 'refname' as the default sort order
The default sort order was '-refname' (reverse lexicographic based on the refname) for FindRefsByOID. The general use case is to use the alphabetical 'refname' (lexicographic based on the refname) sort order, so this MR changes that.

The `git for-each-ref` command that performs the search doesn't need to be passed 'refname' explicitly, as 'refname' is the default for `git for-each-ref`. So the code can be simplified a bit.

Changelog: changed