remote: Try to get hints why FetchInternalRemote tests fail sometimes
The test for `FetchInternalRemote` is still failing from time to time
due to an error when executing the reference-transaction hook. Given
that we don't log errors to stderr in gitaly-hooks but instead to a
hook-specific log file, we have no indicator why the hook is failing
though.

Convert the call to use the `assert` package to continue after
FetchInternalRemote has failed such that we can have a peek at the hooks
log.