gitaly-hooks: Improve logging for the "check" subcommand
The "check" subcommand is different than all the other subcommands of
gitaly-hooks in that it is supposed to be executed by an admin directly
instead of via Git. As such, its logging needs are different than those
of the hooks: it should always print to stdout/stderr directly and not
into the "gitaly_hooks.log". While we do this correctly when using
`log.Fatal()`, the GitLab HTTP client is incorrectly using the hooks log
file as output.

Fix this by using a separate `logrus` instance to log into when creating
the HTTP client. While at it, switch from `log.Fatal()` to explicit
`fmt.Fprintf()` and `os.Exit()` calls -- like this, it becomes harder to
confuse the different logging instances with each other in favor of more
obvious code.