gitaly-hooks: Improve error messages about unexpected args
While we return errors in case we get unexpected arguments in some of
the hooks, these error messages do not state what args we've got. It
would thus be hard to see what went wrong in such a case due to missing
hints.

Improve the error messages by also specifying which args we've got.