gitaly-hooks: Refactor error handling
All hooks return both an error code and an error message, which in most
cases is kind of a duplicate signal in that we mostly return an error
code of `1` in case we've got an error. It's thus easy to get wrong and
may lead to subtle bugs.

Refactor the code to only ever return an error -- if the `main()`
function sees such an error, it knows to log it and return an error code
of `1`. The exception to this is the stdio stream handlers for the
hooks, which also return exit codes of the hooks which have ran. These
now use a special `hookError` error type whihc embeds the returned code
such that we still exit with the correct code.