gitaly-hooks: Get rid of global hook logger
The global hook logger variable is not really required anymore given
that we only set it up and use it in a single location, which is the
`main()` function of gitaly-hooks. Furthermore, tests do not care about
this logger at all except that they need to inject it as a dependency
into `gitlab.NewHTTPClient()`.

Refactor the code to get rid of the global variable.