remote: Fix flaky test caused by unconsumed hook stdin
The test for `FetchInternalRemote()` has by now been flaky for quite a
long time, and various changes which tried to fix it were unsuccessful.
Now that we've amended the test to assert contents of the hook logs we
finally got a hint about what's wrong:

    fetch_internal_remote_test.go:138:
        	Error Trace:	fetch_internal_remote_test.go:138
        	Error:      	Not equal:
        	            	expected: ""
        	            	actual  : "time=\"2021-10-22T06:57:31.170Z\" level=fatal msg=\"error when receiving data for reference-transaction hook: stdin send error: EOF\"\n"

        	            	Diff:
        	            	--- Expected
        	            	+++ Actual
        	            	@@ -1 +1,2 @@
        	            	+time="2021-10-22T06:57:31.170Z" level=fatal msg="error when receiving data for reference-transaction hook: stdin send error: EOF"

        	Test:       	TestFetchInternalRemote_successful

The error happens because the standard input reader which relays
gitaly-hooks' stdin to the hook service is seeing an unexpected EOF
because the remote side (which is the hook service in this case) doesn't
consume it. And this is not much of a surprise: the mocked hook manager
we inject into the server just ignores the stdin reader completely. In
the past, we have held the opinion that it's a bug if the
reference-transaction hook doesn't fully consume stdin because it may
hint either at missing votes or at a vote being computed from partially
read data.

Let's finally fix this bug by just discarding input in our mocked hook
manager.