catfile: Have `Object` implement reader interface directly
The `Object` structure is currently embedding an `objectDataReader`
structure which provides the ability to read the object's data. Like
this, we cannot provide a common `Object` interface that encapsulates
all functions required to inspect Git objects though.

Implement the reader directly on top of the `catfile.Object` to prepare
for the introduction of the `git.Object` interface. This prepares for
also providing the `io.WriterTo` interface for writing object data into
a writer more efficiently.