praefect: Don't modify original message when rewriting during proxying
rewrittenRepositoryMessage function is used by the Coordinator to rewrite
the repository's relative path. The functon signature suggests it doesn't
modify the original request as it returns a marshaled message with the
appropriate fields rewritten. In the process of rewriting, the function
is actually modifiying the original message supplied as a parameter as well.
This is surprising, so this commit modifies the function operate on a clone
of the original message to leave it intact.