praefect: Rewrite additional repository relative path
The Router is already returning the replica path of the additional
repository. This commit updates the Coordinator to rewrite the requests
with the replica path of the additional repository if one is included in
the request. This ensure ObjectPool operations keep functioning as they do
now once Praefect starts generating unique disk paths for repositories.