repocleaner: Don't query fields already known from the database
Repocleaner is currently querying the virtual storage and storage of
a repository from the database. This is unnecessary as the call is
always within a context of a single virtual storage and storage. Given
that, this commit also removes the RepositoryClusterPath struct to simplify
the code. As each of the relative paths in a given storage share the same
virtual storage and relative path, it's not necessary to carry them around
in a struct separately for each relative path.