datastore: Join records by repository ID in StorageCleanup
StorageCleanup was added recently in parallel to the efforts to have
every query join the records by repository ID. As such, it does not
do so yet. This commit updates the query to join by repository ID so
the query will keep working even once Praefect starts generating
relative paths.