Praefect cmd: adding check subcommand
This commit adds a subcommand to run praefect startup checks. This can
be utilized by admins, but can also be scriped (for instanace, in
omnibus) to run before praefect starts up to pervent praefect from
getting into a bad state when there is already an issue with the setup
of the cluster.

Changelog: added