gitpipe: Embed `git.Object` interface into object results
The `CatfileObjectResult` structure currently hosts both an `ObjectInfo`
field and an `ObjectReader`. Ideally, the structure would instead
provide a `catfile.Object` directly, but given that we have to return
different object readers depending on how the results are used this
isn't easily possible.

Fix this issue by introducing a new `git.Object` interface which is
embedded in the structure directly. Like this, the gitpipe code can
return a specific implementation of the code used to synchronize
requesting new objects until the old object has been fully read.