catfile: Use Git to peel tags
When requesting tags via `GetTag()`, then we also try to return
information about the commit an annotated tag points to, if any.
Dereferencing the tag to its tagged object happens in a manual loop
though where we recursive parse the tags until we hit the first non-tag
object or until we have hit a maximum depth of tags to peel. This is
quite inefficient to do and requires us to keep the tag parsing logic
separate from `ParseTag()` itself.

Refactor the code to instead use Git's `^{}` suffix, which will do the
peeling for us to the first non-tag object. Like this we can defer the
heavy lifting to Git, which is a lot more efficient, and eventually
refactor parsing of tags. This should speed up the `UserCreateTag()` and
`GetTagMessages()` RPCs.

Changelog: performance