catfile: Implement parsing of tags on top of struct
In order to parse tags, we use a `bufio.Scanner` to efficiently split up
the commit into lines. The downside of using this interface is that each
scanner has to allocate its own read buffer. In total, when reading many
commits in quick succession, these memory allocations do add up.

Prepare for a fix of this issue by implementing parsing of commits on
top of the newly introduced `parser` structure. Like this, we can
amortize costs of the allocation by reusing the same `bufio.Scanner` for
parsing multiple tags.