catfile: Implement parsing of commits on top of struct
In order to parse commits, we use a `bufio.Reader` to efficiently split
up the commit into lines. The downside of using this interface is that
each newly created reader has to allocate its own read buffer. In total,
when reading many commits in quick succession, these memory allocations
do add up.

Prepare for a fix of this issue by implmeenting parsing of commits on
top of a new `parser` structure. Like this, we can amortize costs of the
allocation by reusing the same `bufio.Reader` for parsing multiple
commits.