replication: Drop 'cancelled' state
The state 'cancelled' is not used in the production
code and the database should not contain any replication
jobs with 'cancelled' state in the replication_queue
table. As the state is not used it makes no sense to
support it. This change removes it from the code and
deletes rows in the replication_queue in that state
(it is safe as praefect has no use case for them).
We can't drop it from the REPLICATION_JOB_STATE type
as we should support compatibility between versions.
The REPLICATION_JOB_STATE type should be cleaned up
in the next release version.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/3859