catfile: Split out trimming of tag messages
The internal `catfile.parseTag()` function takes two parameters to
determine whether the tag message should be trimmed. If set to `true`,
then the message will be trimmed such that it fits into a single gRPC
message and optionally strip all trailing newlines. And by default,
`ParseTag()` will trim messages like this, which is hidden away from the
caller given that this trimming cannot be controlled from that function.

Parsing tags and trimming resulting messages doesn't feel like the right
thing to mix with each other: it's unexpected to get a parsed tag which
doesn't match actual contents of the tag, and it complicates the parsing
logic itself. We still cannot get rid of the truncation altogether, but
as it turns out there are only so many callsites which do in fact want
this parsing. As such it is easy to split out this logic into a separate
function.

Extract this functionality into a separate `TrimTagMessage()` function.
Given that all callsites either set both trimming parameters to `true`
or `false`, this function doesn't have any parameters at all. This is
both a lot more explicit at the callsite, doesn't silently truncate
messages and allows us to more easily refactor the tag parsing.