catfile: Merge splitting and parsing of tags
The current tag parser consists of two parts: first we split the tag
into headers and message, and then we parse the header. This split makes
it hard to optimize reading of tag data, and furthermore it's not
required anymore after we only have a single caller of the splitting.

Merge both parts into `parseTag()` to prepare for a rewrite of it in
order to improve memory efficiency. Note that the function is already
converted to be a receiver function on fop of the `parser` structure
while at it.