catfile: Rewrite tag parser to amortize memory allocations
The tag parser is using a `bufio.Scanner` to scan object data and thus
construct the tag structure. This interface is allocating an internal
buffer to buffer writes, and given that we're allocating a new scanner
for each object we're about to scan this cost quickly adds up.
Furthermore, scanners are restricted with regards to their maximum line
length they can process, which may easily be exceeded with specially
crafted inputs.

Refactor the parser to instead use a `bufio.Reader` similar, where we
follow the same logic to parse tags as we do when parsing commits. This
allows us to easily reuse the same reader across different calls to
`parseTag()` and thus to amortize costs of the allocation. Benchmarks
before this change:

    BenchmarkFindAllTags/FindAllTags-16   6   191726245 ns/op   54665286 B/op   148914 allocs/op

And after this change:

    BenchmarkFindAllTags/FindAllTags-16   6   171534539 ns/op   30851421 B/op   139056 allocs/op

This shows a nice decrease of allocated bytes of almost 45%.

Changelog: performance