testhelper: Change `ModifyEnvironment()` to unset envvars on empty value
In a subsequent commit, we'll want to modernize some tests to make
better use of the testhelper package, where some of these tests need to
unset certain environment variables if they're set. Ideally we would use
our `ModifyEnvironment()` test helper for this, but it can only set an
envvar to a new value as compared to unsetting it altogether.

Change the function to unset envvars if it is given an empty string as
value. While this is a backwards-incompatible change, no tests seem to
currently make use of empty values, so it shouldn't be much of an issue
to change it like this.