test: Fix flaky TestStorageCleanup_AcquireNextStorage
The verification of the value filled for the TriggeredAt
is flaky as precision of the timestamp returned from the
database is less than used in code. That is why value of
the start variable often is greater than one contained in
the TriggeredAt column. As this check is not mandatory
we just skip it and verify the value is set.