praefect: Don't query for replica path in NodeManagerRouter
Rails tests are stuck with using NodeManagerRouter and the legacy
election strategies due to not using a database. As there's no
database, the NodeManagerRouter is used with a stubbed out RepositoryStore.
This stub doesn't know anything about the replica paths the repositories
have. As the NodeManagerRouter and the election strategies are unsupported
and don't need to support the Praefect generated relative paths, let's
just use the relative path as the replica path to fix the tests.