supervisor: Fix RSS monitor retriggering too fast
When we supervise a running process, then regularly poll consumed memory
by the process such that the user of the supervisor can kill the process
if it blows up too much. The intent is to do this once every 15 seconds,
but because we create the timer ticker before we have actually received
information about the process to be monitored it can cause us to poll
the process twice in a row when it takes more than 15 seconds to get
that information.

Fix the issue by resetting the timer before doing the select such that
we always wait for 15 seconds. Note that this also converts us to use a
`helper.TimerTicker`: the benefit is that its reset function knows to
drain already-pending events.

Changelog: fixed