datastore: Consider replicas pending rename outdated
GetConsistentStorages has a regression where a replica pending rename
is considered up to date. Requests routed to the replica would fail as
the replica is still stored at the old path. Previously the records were
joined by the (virtual_storage, relative_path) which naturally considered
the rename pending replicas outdated. Until the relative_path is completely
removed from storage_repositories, this commit updates the query to ensure
the replica's relative path is the same as the repository's. This indicates
the rename operation has been applied.