reconciler: Consider replicas pending rename outdated
Reconciler is scheduling update jobs to replicas that are pending
renames. This began happening when we updated reconciler to join
the records by repository id in 2b19d8c5. As the renames previously
disjointed the records in `repositories` and `storage_repositories`
due to the `relative_path` changing, the reconciler didn't target
replicas that were pending rename. This commit changes the query to
only use replicas which are not pending a rename to act as a source
node for replication. Combined with a previous commit to update the
relative path in SetGeneration if necessary, this will restore the
same behavior the reconciler had earlier with renames; never using
replicas pending rename as a source but possibly creating a new
replica.