ci: Remove unused files from cache
Back when we introduced the libgit2 and git build targets, we generated
a `Makefile.sha256` file to see whether contents of the Makefile have
changed and thus whether we need to rebuild these targets. This has
since been replaced with a much finer-grained mechanism which records
the version of each of our build targets separately such that only those
targets are rebuild which have actually changed. Our CI cache still
refers to the old file path though, which is nowadays unused.

Remove references to `Makefile.sha256`.