gitpipe: Replace manual for-each-ref decoding
Previously ref parsing was done in place. This replaces the parsing with
a shared implementation.

To enable the git output to be decoded with the same code the format was
changed to a consistent NUL byte separator. `PEELED` had to be
lower-cased as git was interpreting %00PEELED as an escape code.