praefect: Configurable dial timeout
The function subCmdDial uses grpc.WithBlock() option
to connect with the remote gRPC service. It means
the DialContext will be blocked until connection is
set. If connection can't be established the lib
attempts to dial for some time that is configurable
by providing context with timeout set. The default
timeout we are using is 30 sec. Exactly this time
will be wasted in the 'one of gitalies is out of service'
test case of the TestRemoveRepository_Exec test.
To mitigate this problem the subCmdDial function is
extended with context and timeout parameters. Now we
override timeout value in the test to speed it up.