repository: Stop calling CreateRepository to create hooks symlink
Back in the days, we used to set up Git hooks by creating a symlink to
our directory containing the actual hook scripts. This has long be
replaced by a much more robust mechanism which injects "core.hooksPath"
into Git processes to point to that directory instead. Since then, we
don't really need to setup hooks on-disk anymore, except if an admin
wants to create per-project server-side hooks.

Back when we used to create the symlink though we had to make sure that
all repos which got newly initialized did in fact have the symlink. To
do so, we used to call `CreateRepository()` after having set up repos,
which created the symlink for us. Given that the call is idempotent, it
was totally fine if the repository was already initialized: git-init(1)
simply re-copies the templates into the repository and creates anything
that's missing.

Nowadays this call is superfluous though: we have just created the repo,
we don't want to set up a symlink and thus it ultimately doesn't serve
any purpose at all anymore. So let's get rid of this useless step and
drop calls to `CreateRepository()` in `CreateRepositoryFromBundle()`,
`CreateRepositoryFromURL()` and `CreateFork()` and start to assert that
the hook directory is a directory, like it would be created by Git.

Note that this also removes the transactional voting step as performed
by CreateRepository. The voting in there is broken anyway given that it
doesn't guarantee any atomicity anyway. This is being remodelled in
another MR in the future which updates creation of repos to become
atomic, and dropping the CreateRepository call now goes into the same
direction.