ci: Use non-merged result SHA for QA pipeline
When the merged result pipeline is triggered the child pipeline clones
Gitaly. But it does not fetch the merged result revision, because
there's no branch referencing it. Ideally we should change the
downstream pipeline so it fetches that revision [1], but for now just
checkout the HEAD revision on the merge request source branch.

Now because we don't know whether the pipeline is triggered on a merged
result, merge train, or default branch, it might be that
CI_MERGE_REQUEST_SOURCE_BRANCH_SHA is not defined. To overcome this the
job is rewritten to use the rules definitions. This allows us to set
variables depending on some conditions.

This change needs `allow_failure: true` because it would otherwise stop
the pipeline [2].

1. https://gitlab.com/gitlab-org/omnibus-gitlab/-/issues/6482
2. https://docs.gitlab.com/ee/ci/yaml/index.html#rulesallow_failure

Issue: https://gitlab.com/gitlab-org/gitaly/-/issues/3852