gittest: Convert `ExecStream()` to take a configuration
The `gittest.ExecStream()` helper takes an additional reader which gets
used as the spawned command's standard input. In order to make more use
of this helper though we'll need to add additional arguments, which
would eventually become quite unwieldy to use.

Convert the function to instead take the reader via a config struct such
that it's more readily extensible.