ssh: Convert clones to avoid executing Git directly
The SSH upload-pack tests execute Git directly. This will stop working
as soon as we migrate to use a bundled version of Git given that we
wouldn't be able to locate git-upload-pack(1) anymore.

Convert the tests to instead use a `git.ExecCommandFactory()` to prepare
the clone commands and execute them to be prepared for this change.