repository: Do not execute Git binary directly
The testcase for cloning from a redirecting server uses `exec.Cmd` to
spawn the Git executable directly. While this works now, it will stop
working as soon as we bundle Git as a binary given that we wouldn't be
able to find the git-remote-http helper anymore.

Prepare for this change by using a `git.ExecCommandFactory` instead to
spawn the command.