gitaly-hooks: Do not execute Git binary directly
One of our gitaly-hooks tests which exercises the pack-objects hook is
using `os.Exec()` directly to spawn a Git command. This will stop
working when we migrate to using a bundled Git version given that we
wouldn't be able to locate Git's exec prefix anymore.

While this testcase is a good candidate to use `gittest.ExecOpts()`,
what is still missing is the ability to pass additional environment
variables to the spawned command. Extend the config struct to allow for
this and convert the test.