gittest: Convert HTTP server to not spawn Git directly
The test HTTP server set up by the gittest package is using `exec.Cmd`
to spawn a Git executable directly. This will stop working when we
migrate to a bundled Git version because we'd be unable to locate the
git-http-backend helper.

The HTTP server requires the ability to set up stdout of the command
such that it can serve as a CGI handler. Extend the `gittest.ExecOpts`
helper to support setup of stdout and convert the HTTP server setup to
use it.