catfile: Fix correlation IDs for tracing spans
When creating a new tracing span in the catfile package, then we need to
create those spans for two different contexts: the usual per-RPC context
and the decorrelated context that's stored as part of the cache. In
order to associate spans created in the decorrelated context with their
originating RPC, we tag the span with the correlation ID of the per-RPC
context. But because we pass arguments to `startSpan()` in the wrong
order, we accidentally did it the other way round: the per-RPC context
ends up with the correlation ID of the cached context.

Fix this bug by switching the order in which we pass the contexts.