catfile: Fix double-accounting of traces for uncached processes
In the catfile cache, we need to discern processes which are cached and
those which aren't cached, e.g. because they are not created in a
cacheable context. So while a catfile process always has two different
contexts with the per-RPC context and the decorrelated context stored in
the cache, these contexts may indeed be the same. But in fact, our
tracing infrastructure in that package will always create tracing
spans for both contexts, even if they're the same, which thus leads to
double-accounting of each request.

Fix the bug by only creating the secondary span in case it's different
from the primary per-RPC context.