catfile: Disentangle object reader's and object's closed status
Same as with the preceding commit, the object reaches into the parent to
determine whether the object reade is closed or not. This comlicates
locking and is unclean design-wise given that concerns of the child
should be disentangled from its parent.

Fix this design issue by keeping track of the closed status in object
reader and object separately. This allows us to now remove the reference
to the parent reader completely.