commit: Fix discarding of tree entries
When retrieving tree entries via the tree entries helper, then we
request an object, parse it and then recursively parse its children in
case it is a tree object. When the object is not a tree, then we discard
its contents via a deferred call and exit early. Otherwise, if it is a
tree, we'd still try to discard the tree object _after_ we have parsed
all of its children. This is invalid though given that objects returned
by the catfile package cannot be read in arbitrary order, but instead
they must always be fully read before trying to read the next object.
Discarding contents at the end thus cannot work as expected.

Fix the bug by explicitly discarding the object in case it's not a tree.
In the other case, where it is a tree, we'll pass the reader to
`extractEntryInfoFromTreeData()`: this function either returns an error
in case we're unable to parse the tree, or otherwise it has fully
consumed the tree object. Discarding it is thus not necessary in that
case.

Changelog: fixed