catfile: Fix error messages on cached process type mismatches
When retrieving processes from the catfile process cache, then we must
assert that their type matches the expected type we wanted to retrieve.
In case they don't match, we return an error detailing which type we've
got instead. As it turns out, the error message is wrong though: it will
always see a `nil` pointer given that we pass the object which we failed
to cast to, instead of the object we want to cast.

Fix the error message by passing the correct object.